Inheritance Day 2

+++ src/Inheritance/Quadrilateral.java
@@ -1,10 +1,10 @@
 package Inheritance;
 
 public class Quadrilateral {
-	private int s1;
-	private int s2;
-	private int s3;
-	private int s4;
+	private double s1;
+	private double s2;
+	private double s3;
+	private double s4;
 
 	public Quadrilateral() {
 		s1 = 5;
@@ -13,35 +13,35 @@ public class Quadrilateral {
 		s4 = 8;
 	}
 
-	public Quadrilateral(int xS1, int xS2, int xS3, int xS4) {
+	public Quadrilateral(double xS1, double xS2, double xS3, double xS4) {
 		s1 = xS1;
 		s2 = xS2;
 		s3 = xS3;
 		s4 = xS4;
 
 	}
-	public void setS1(int xS1){
+	public void setS1(double xS1){
 		s1 = xS1;
 	}
-	public void setS2(int xS2){
+	public void setS2(double xS2){
 		s2 = xS2;
 	}
-	public void setS3(int xS3){
+	public void setS3(double xS3){
 		s3 = xS3;
 	}
-	public void setS4(int xS4){
+	public void setS4(double xS4){
 		s4 = xS4;
 	}
-	public int getS1(){
+	public double getS1(){
 		return s1;
 	}
-	public int getS2(){
+	public double getS2(){
 		return s2;
 	}
-	public int getS3(){
+	public double getS3(){
 		return s3;
 	}
-	public int getS4(){
+	public double getS4(){
 		return s4;
 	}
 	public double perimeter(){
